Update the game page for SR2014.
This changes the tense used in the 2013 piece to be historic and
moves that down under a new heading. Add information about this
year's game, adding useful links to other pages (the expectation
is that these will remain with the 'current' year as time passes).

This includes an image of the slots/zones construction, exported
from rules.git/images/slots-zones.svg@27cb648d472132.

Change-Id: I2bc417befc26996f40fbc451da81a60428d227d9