Add instructions on using the tablet web app
The tablet page currently doesn't provide any instruction on the use of
the web app. This patch adds instructions on using the setup and log
pages within the web app.

Change-Id: I23e4d7df7c3534b00a4ef7d6c48eb990d4bfe472

--

This commit is a reapplication of f1fe2e8cb271e91dfbe6873d97058f736d11dcae
from srweb, which appears to have been lost in the move to this new repo.

This commit doesn't include the CSS change from the original as it's not
clear where to apply that change, and since it appears unneccesary due to
the way images are handled here at the moment.