Rewrite the tablet page from a general perspective
We provide the tablets in the kit in case the team does not have access
to their own computing device with WiFi. However, the vast majority of
teams do have access to such devices so the docs should be written from
that perspective.

This rewrite removes references to the tablet and makes the docs
applicable to all devices.