Change sidebar highlights to allow an appended '/'
Previously if a url in _data/sidebar_tree.yaml had a trailing '/'
character, the sidebar highlighting would not match the page url
and therefore wouldn't highlight the current page.

This change adds some robustness to the matching to strip any
extra '/' characters, allowing us to append them to urls
in the data file again.