Avoid invalid HTTPS certificat error on xgoat by going direct
This changes the links to our parts to avoid the convenient but
unneccesary redirection via www.xgoat.com, which has recently
developed an HTTPS certificate issue (notably that its cert is
for 'xgoat.com' and not the subdomain 'www.xgoat.com' which we're
using).

While that issue could easily be worked around by using a more
direct xgoat link, we might as well just go direct to the supplier
links as they're also stable and remove a dependency.