Fully replace the safety regs with those from the SR2020.5 rules
The regs which were here appear date to a previous version of the
kit and are incompatible with the current version. Surprisingly
this appears to have already been the case when they were added
to this repo, though that appears to have been missed at the time.

This commit introduces the regulations as they were in the SR2020.5
tag of the rules (the last published version before the move to
HTML format).

https://github.com/srobo/rules/commit/00e4ede22f21b2c8e99d5aa5928fe60b4e1d2cad

Fixes https://github.com/srobo/docs/issues/324