Document how to use the LEDs on the KCH HAT
I've opted to name this document after the LEDs it controls,
rather than the board where those LEDs are hosted, to follow
the naming pattern of the "motors", "servos" (etc.) docs which
do the same for the other boards.

This is based on the current state of the API integration in
https://github.com/srobo/sr-robot3/pull/60, so may not be quite
how this is intended to function. Hopefully there's a useful
starting point here though.

Fixes https://github.com/srobo/docs/issues/358