Enact the third rules update for SR2023
This changes the sizes of the markers and the Bronze & Silver tokens to
enable the markers to actually fit on the tokens. This is due to the white
border around the marker being necessary for the vision system to work,
but being outside the specified size of the markers.

This is built from:
- https://github.com/srobo/rules/commit/77d0dc8e9345654bac2b51430fdd648e2f6e29a0
- https://app.circleci.com/pipelines/github/srobo/rules/247/workflows/dfd11c9d-706c-4ff9-a346-54316dc4cc33/jobs/316/artifacts