Enact the fourth rules update for SR2023
This specifies the mass of the Bronze tokens to enable teams to
more easily work on their robot designs. The mass of the Silver
tokens can be inferred from this easily. The mass of the Gold tokens
remains undefined, though we have provided guidance on that via
Discord.

This is built from:
- https://github.com/srobo/rules/commit/b9923baef45c72584a677480bd85ad8d092d4420
- https://app.circleci.com/pipelines/github/srobo/rules/253/workflows/2a1b8b66-4aea-436c-87e3-b2969d41804c/jobs/321/artifacts