Update the vision docs to match reality
Even before the changes in the Dec 1st kit update these docs were
wildly different to how the vision system actually behaved.
This update brings these to parity with the default behaviour of
the kit (and soon the simulator).

Ideally we would have some diagrams to show how the axes work,
however the diagrams we have are substantially incorrect and will
need replacing. Dropping them clarifies that work is needed before
they can be brought back (though recreating them from scratch
might be easier as we don't seem to have the source files).