Move the SR2023 rules to the archive
The arena image was constructed by manually adjusting the diagram
from the rules to remove the dimensions and other artefacts, then
exporting as PNG via inkscape.

Where previously we've replaced the content of the rules index page
entirely at this point in the year, it feels like having a conditional
(even a constant one) will make this process slightly simpler in future.