Fix confused usage of `input` in the Python tutorial
From the context it looks like this was previously comparing
Python 2's `raw_input` and `input` functions, and that this was
missed during the upgrade to Python 3. Drop the comparison and
explicitly parse the inputs to numbers to let the reader focus
on the task at hand (computing the average).

Fixes https://github.com/srobo/docs/issues/533