Clarify the nature of the Docker alternative and add links
Importantly this pushes this option to being secondary and
clarifies which of the build steps it replaces.

Note that this also removes the instructions on how to use Docker.
If users are not already familiar with Docker & Docker Compose
then they are equally not equipped to judge the security posture
that it presents.