Add general information about our events
This mostly serves as a place to document that we don't require
the primary contact to be the person who brings a team to our
events, which is something that we're asked every year. We've
pretty much always had the same policy, however tend to need to
re-evaluate it every time it's asked as we haven't written it down.

Fixes https://github.com/srobo/tasks/issues/1475