Drop 'neat' gem
This is a SASS grid package, though as far as I can tell we're not
using it. The [docs](https://neat.bourbon.io/) suggest we'd need
to `@import "neat"` in our SASS files to be using it, which we're
not. A cursory browse around the site also seems to work fine.